# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Toodl Expense Splitter API",
    "description": "API for managing expense groups, splitting bills, and tracking who owes what. Perfect for shared expenses like trips, dinners, rent, and more.",
    "version": "1.0.0",
    "contact": {
      "name": "Toodl Support",
      "url": "https://toodl.co"
    }
  },
  "servers": [
    {
      "url": "https://toodl.co/api",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
  "paths": {
    "/groups": {
      "post": {
        "summary": "Create a new expense group",
        "description": "Create a new group for splitting expenses. The creator becomes the first member automatically.",
        "operationId": "createGroup",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "creatorName"],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Name of the expense group (e.g., 'Weekend Trip', 'Apartment Rent')",
                    "example": "Weekend Trip"
                  },
                  "creatorName": {
                    "type": "string",
                    "description": "Name of the person creating the group",
                    "example": "Alice"
                  },
                  "members": {
                    "type": "array",
                    "description": "Additional members to add to the group",
                    "items": {
                      "type": "object",
                      "properties": {
                        "firstName": {
                          "type": "string",
                          "description": "First name of the member",
                          "example": "Bob"
                        },
                        "email": {
                          "type": "string",
                          "description": "Email address of the member (optional)",
                          "example": "bob@example.com"
                        }
                      },
                      "required": ["firstName"]
                    }
                  },
                  "currency": {
                    "type": "string",
                    "description": "Currency code for the group",
                    "enum": ["USD", "EUR", "INR", "GBP", "CAD", "AUD"],
                    "default": "USD",
                    "example": "USD"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Group created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "groupId": { "type": "string" },
                    "groupName": { "type": "string" },
                    "members": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "firstName": { "type": "string" },
                          "email": { "type": "string" },
                          "authProvider": { "type": "string" }
                        }
                      }
                    },
                    "currency": { "type": "string" },
                    "shareLink": { "type": "string" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/groups/{groupId}": {
      "get": {
        "summary": "Get group details",
        "description": "Retrieve all information about a specific group including members, expenses, and summary statistics.",
        "operationId": "getGroup",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Unique identifier of the group"
          }
        ],
        "responses": {
          "200": {
            "description": "Group details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "group": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "name": { "type": "string" },
                        "members": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": { "type": "string" },
                              "firstName": { "type": "string" },
                              "email": { "type": "string" },
                              "authProvider": { "type": "string" }
                            }
                          }
                        },
                        "expenses": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": { "type": "string" },
                              "description": { "type": "string" },
                              "amount": { "type": "number" },
                              "paidBy": { "type": "string" },
                              "splits": { "type": "object" },
                              "createdAt": { "type": "string" }
                            }
                          }
                        },
                        "currency": { "type": "string" },
                        "createdAt": { "type": "string" },
                        "lastUpdated": { "type": "string" }
                      }
                    },
                    "summary": {
                      "type": "object",
                      "properties": {
                        "totalExpenses": { "type": "number" },
                        "totalExpensesFormatted": { "type": "string" },
                        "memberCount": { "type": "number" },
                        "expenseCount": { "type": "number" }
                      }
                    },
                    "shareLink": { "type": "string" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Group not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}